fixes #357
