+++ _posts/v.3.0.0.md
@@ -7,6 +7,7 @@ YASP is proud to present our latest major release, v3.0.0!
 ### Features
 * Teamfights.  Get a breakdown on the teamfights that occurred during a match!
 * Objectives.  See when towers, Roshan, barracks, and the Ancient fell.
+* MMR Leaderboard.  See who the top YASP players are!
 * Winrate by hour/day.  See when you play at your best!
 * Improved largest hit.  See what you hit with, and who you hit hard in each game!
 * General match querying (from Matches page)  Run advanced querying on the general match pool.  Wonder how well Terrorblade and Broodmother perform together?  Now you can find out!
@@ -14,7 +15,8 @@ YASP is proud to present our latest major release, v3.0.0!
 
 ### UI
 * Reorganized player pages.  You can now do match querying directly from your home page.  Aggregations remain on the Trends tab.  The Activity calendar has been moved back to the main page.
-
+* Reordered items in navbar
+* 
 ### Bugfixes
 * Some matches may have been missing chat messages.  We've changed our storage format and should now capture chat events that we couldnt link to a player.
 
@@ -22,3 +24,4 @@ YASP is proud to present our latest major release, v3.0.0!
 * Parsed data v7.  Updated format to support more fields.
 * Improved frontend dependency management (better maintainability)
 * Modified process for retrieving parsed data from DB.  Should lead to better load times for Trends tabs.
+* DB data migration.  Converted data from old formats to avoid doing it on-the-fly.
