+++ sources.json
@@ -14,6 +14,12 @@
     "5": "Never Connected",
     "6": "Never Connected (Timeout)"
   },
+  "skillshots": {
+    "mirana_arrow": true,
+    "pudge_meat_hook": true,
+    "rattletrap_hookshot": true,
+    "invoker_sun_strike": true
+  },
   "gold_reasons": {
     "0": {
       "name": "Other",
