+++ views/base_layout.jade
@@ -57,7 +57,7 @@ html
       link(rel="stylesheet", href='//cdn.jsdelivr.net/select2/latest/select2-bootstrap.css')
       script(type='text/javascript' src='//cdn.jsdelivr.net/select2/latest/select2.min.js')
       //d3
-      script(type='text/javascript' src="//cdn.jsdelivr.net/d3js/3.5.5/d3.min.js")
+      script(type='text/javascript' src="//cdn.jsdelivr.net/d3js/latest/d3.min.js")
       //c3
       script(type='text/javascript' src='//cdn.rawgit.com/masayuki0812/c3/0.4.10/c3.min.js')
       link(rel='stylesheet', href='//cdn.rawgit.com/masayuki0812/c3/0.4.10/c3.min.css')
