+++ views/navbar_content.jade
@@ -11,7 +11,7 @@
       li
         a(href='/matches') Matches
       li
-        a(href='/status') Status
+        a(href='/ratings') Ratings
       li
         a(href="/request") Request
       li
@@ -20,6 +20,8 @@
         a(href='/page/1') Blog      
       li
         a(href='/carry') Help Carry!
+      li
+        a(href='/status') Status
     ul.nav.navbar-nav.navbar-right
       if user
         li
