+++ views/ratings.jade
@@ -1,0 +1,20 @@
+extend layout
+
+block content
+  pre=JSON.stringify(ratings)
+  h1 Ratings
+  table.table#ratings
+    thead
+      tr
+        th Name
+        th Solo
+        th Party
+        th Updated
+    each rating in ratings
+      tr
+        td
+          img.user(src=rating.avatar)
+          a(href='/players/#{rating.account_id}')=rating.personaname || rating.account_id
+        td= rating.soloCompetitiveRank
+        td= rating.competitiveRank
+        td= moment(rating.time).fromNow()
