+++ views/request.jade
@@ -3,7 +3,11 @@ extend layout
 block content
   .row
     .col-md-4
-      h1 Match Requests
+      h1 Match Requests 
+      ul
+        li Enter a match ID
+        li Prove your humanity
+        li We'll try to parse it for you
       div#request
         .form-group
           input#match_id.form-control(type='text' name='match_id' placeholder='Match ID')
